RP-366 Create Sandbox API/Service Integration change - updated API to wrap the common properties.
