+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/model/ApplicationSandboxConfiguration.java
@@ -1,23 +1,13 @@
 package org.opentestsystem.rdw.admin.model;
 
-import org.opentestsystem.rdw.multitenant.Tenant;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesTenant;
 import org.opentestsystem.rdw.reporting.common.configuration.ReportingSystemPropertiesImpl;
 
-import java.util.Set;
-
 /**
- * Used to manage a configuration properties for the Sandbox service API
+ * The Common Sandbox properties
  */
 public class ApplicationSandboxConfiguration {
-    /**
-     * Tenant used to create sandbox, and where sandbox exists
-     */
-    private Tenant tenant;
-    /**
-     * Collection of dataSets preconfigured for this tenant
-     */
-    private Set<DataSet> dataSets;
     /**
      * Default Reporting system properties
      */
@@ -25,8 +15,11 @@ public class ApplicationSandboxConfiguration {
     /**
      * Default aggregate reporting properties
      */
+    @JsonProperty("aggregator")
+    //todo DeserializationFeature.ACCEPT_SINGLE_VALUE_AS_ARRAY`
     private AggregateReportingPropertiesTenant aggregateReportingPropertiesTenant;
 
+    ApplicationSandboxConfiguration () {}
 
     public ReportingSystemPropertiesImpl getReporting() {
         return reporting;
@@ -34,22 +27,6 @@ public class ApplicationSandboxConfiguration {
 
     public void setReporting(final ReportingSystemPropertiesImpl reporting) {
         this.reporting = reporting;
-    }
-
-    public Tenant getTenant() {
-        return tenant;
-    }
-
-    public void setTenant(final Tenant tenant) {
-        this.tenant = tenant;
-    }
-
-    public Set<DataSet> getDataSets() {
-        return dataSets;
-    }
-
-    public void setDataSets(final Set<DataSet> dataSets) {
-        this.dataSets = dataSets;
     }
 
     public AggregateReportingPropertiesTenant getAggregateReportingPropertiesTenant() {
