+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/model/ApplicationSandboxConfigurations.java
@@ -6,15 +6,15 @@ import java.util.Set;
  * All of the application Default properties for a Sandbox
  */
 public class ApplicationSandboxConfigurations {
-    private ApplicationSandboxConfiguration applicationSandboxConfiguration;
+    private SandboxConfigurationPackage sandboxConfigurationPackage;
     private Set<SandboxConfiguration> sandboxes;
 
-    public ApplicationSandboxConfiguration getApplicationSandboxConfiguration() {
-        return applicationSandboxConfiguration;
+    public SandboxConfigurationPackage getSandboxConfigurationPackage() {
+        return sandboxConfigurationPackage;
     }
 
-    public void setApplicationSandboxConfiguration(final ApplicationSandboxConfiguration applicationSandboxConfiguration) {
-        this.applicationSandboxConfiguration = applicationSandboxConfiguration;
+    public void setSandboxConfigurationPackage(final SandboxConfigurationPackage sandboxConfigurationPackage) {
+        this.sandboxConfigurationPackage = sandboxConfigurationPackage;
     }
 
     public Set<SandboxConfiguration> getSandboxes() {
