+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/model/SandboxConfiguration.java
@@ -12,9 +12,9 @@ import java.util.Map;
  */
 public class SandboxConfiguration {
     private Sandbox sandbox;
-    private ReportingSystemPropertiesImpl reporting;
-    @JsonProperty("aggregator")
-    private AggregateReportingPropertiesTenant aggregateReportingPropertiesTenant;
+
+    private ApplicationSandboxConfiguration applicationSandboxConfiguration;
+
     private Map<String, String> localization = new LinkedHashMap<>();
 
     public Sandbox getSandbox() {
@@ -25,27 +25,19 @@ public class SandboxConfiguration {
         this.sandbox = sandbox;
     }
 
-    public ReportingSystemPropertiesImpl getReporting() {
-        return reporting;
-    }
-
-    public void setReporting(final ReportingSystemPropertiesImpl reporting) {
-        this.reporting = reporting;
-    }
-
-    public AggregateReportingPropertiesTenant getAggregateReportingPropertiesTenant() {
-        return aggregateReportingPropertiesTenant;
-    }
-
-    public void setAggregateReportingPropertiesTenant(final AggregateReportingPropertiesTenant aggregateReportingPropertiesTenant) {
-        this.aggregateReportingPropertiesTenant = aggregateReportingPropertiesTenant;
-    }
-
     public Map<String, String> getLocalization() {
         return localization;
     }
 
     public void setLocalization(final Map<String, String> localization) {
         this.localization = localization;
+    }
+
+    public ApplicationSandboxConfiguration getApplicationSandboxConfiguration() {
+        return applicationSandboxConfiguration;
+    }
+
+    public void setApplicationSandboxConfiguration(final ApplicationSandboxConfiguration applicationSandboxConfiguration) {
+        this.applicationSandboxConfiguration = applicationSandboxConfiguration;
     }
 }
