+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/model/SandboxConfigurationPackage.java
@@ -1,0 +1,49 @@
+package org.opentestsystem.rdw.admin.model;
+
+import org.opentestsystem.rdw.multitenant.Tenant;
+
+import java.util.Set;
+
+/**
+ * Used to manage a configuration properties for the Sandbox service API
+ */
+public class SandboxConfigurationPackage {
+    /**
+     * Tenant used to create sandbox, and where sandbox exists
+     */
+    private Tenant tenant;
+    /**
+     * Collection of dataSets preconfigured for this tenant
+     */
+    private Set<DataSet> dataSets;
+    /**
+     * The common configuration properties
+     */
+    private ApplicationSandboxConfiguration applicationSandboxConfiguration;
+
+    SandboxConfigurationPackage() {}
+
+    public Tenant getTenant() {
+        return tenant;
+    }
+
+    public void setTenant(final Tenant tenant) {
+        this.tenant = tenant;
+    }
+
+    public Set<DataSet> getDataSets() {
+        return dataSets;
+    }
+
+    public void setDataSets(final Set<DataSet> dataSets) {
+        this.dataSets = dataSets;
+    }
+
+    public ApplicationSandboxConfiguration getApplicationSandboxConfiguration() {
+        return applicationSandboxConfiguration;
+    }
+
+    public void setApplicationSandboxConfiguration(final ApplicationSandboxConfiguration applicationSandboxConfiguration) {
+        this.applicationSandboxConfiguration = applicationSandboxConfiguration;
+    }
+}
