+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/model/SandboxConfigurationStubs.java
@@ -8,19 +8,17 @@ import java.util.Map;
 import java.util.Set;
 
 /**
- * ApplicationSandboxConfiguration and Sandbox configurations used for testing
+ * SandboxConfigurationPackage and Sandbox configurations used for testing
  * And to serve up mock data
  * TODO This will be moved to test once full sandbox configurations are managed
  */
 public class SandboxConfigurationStubs {
     // Default Tenant will be California
-    public static ApplicationSandboxConfiguration getApplicationSandboxConfigurationDefault() {
-        ApplicationSandboxConfiguration applicationSandboxConfiguration = new ApplicationSandboxConfiguration();
-        applicationSandboxConfiguration.setTenant(TenantConfigurationStubs.getTenantCA());
-        applicationSandboxConfiguration.setDataSets(getDataSets());
-        applicationSandboxConfiguration.setReporting(TenantConfigurationStubs.getReporting());
-        applicationSandboxConfiguration.setAggregateReportingPropertiesTenant(getAggregrateReportingProperties());
-        return applicationSandboxConfiguration;
+    public static SandboxConfigurationPackage getApplicationSandboxConfigurationDefault() {
+        SandboxConfigurationPackage sandboxConfigurationPackage = new SandboxConfigurationPackage();
+        sandboxConfigurationPackage.setTenant(TenantConfigurationStubs.getTenantCA());
+        sandboxConfigurationPackage.setDataSets(getDataSets());
+        return sandboxConfigurationPackage;
     }
 
     public static Set<DataSet> getDataSets() {
@@ -45,18 +43,22 @@ public class SandboxConfigurationStubs {
     public static SandboxConfiguration getCaInterimSandboxConfiguration() {
         SandboxConfiguration sandboxConfiguration = new SandboxConfiguration();
         sandboxConfiguration.setSandbox(getSandboxCAInfo());
-        sandboxConfiguration.setReporting(TenantConfigurationStubs.getReporting());
+        sandboxConfiguration.setApplicationSandboxConfiguration(getApplicationSandboxConfiguration());
         sandboxConfiguration.setLocalization(getCaLocalization());
-        sandboxConfiguration.setAggregateReportingPropertiesTenant(getAggregrateReportingProperties());
         return sandboxConfiguration;
     }
 
+    private static ApplicationSandboxConfiguration getApplicationSandboxConfiguration() {
+        ApplicationSandboxConfiguration applicationSandboxConfiguration = new ApplicationSandboxConfiguration();
+        applicationSandboxConfiguration.setReporting(TenantConfigurationStubs.getReporting());
+        applicationSandboxConfiguration.setAggregateReportingPropertiesTenant(getAggregrateReportingProperties());
+        return applicationSandboxConfiguration;
+    }
     public static SandboxConfiguration getMiInterimSandboxConfiguration() {
         SandboxConfiguration sandboxConfiguration = new SandboxConfiguration();
         sandboxConfiguration.setSandbox(getSandboxMIInfo());
-        sandboxConfiguration.setReporting(TenantConfigurationStubs.getReporting());
+        sandboxConfiguration.setApplicationSandboxConfiguration(getApplicationSandboxConfiguration());
         sandboxConfiguration.setLocalization(getCaLocalization());
-        sandboxConfiguration.setAggregateReportingPropertiesTenant(getAggregrateReportingProperties());
         return sandboxConfiguration;
     }
 
@@ -98,6 +100,7 @@ public class SandboxConfigurationStubs {
         assessmentTypes.add("iab");
         aggregateReportingProperties.setAssessmentTypes(assessmentTypes);
         Set<String> assessments = new HashSet<>();
+        assessmentTypes.add("ica"); // todo change back to one once deserializer for fields are in place
         assessmentTypes.add("sum");
         aggregateReportingProperties.setStateAggregateAssessmentTypes(assessments);
         Set<String> stateAssessments = new HashSet<>();
