+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/service/impl/DefaultSandboxService.java
@@ -36,7 +36,7 @@ public class DefaultSandboxService implements SandboxService {
     public ApplicationSandboxConfigurations getAll() {
         ApplicationSandboxConfigurations applicationSandboxConfigurations = new ApplicationSandboxConfigurations();
         applicationSandboxConfigurations.setSandboxes(inMemorySandboxConfigurationRepository.getAll());
-        applicationSandboxConfigurations.setApplicationSandboxConfiguration(SandboxConfigurationStubs.getApplicationSandboxConfigurationDefault());
+        applicationSandboxConfigurations.setSandboxConfigurationPackage(SandboxConfigurationStubs.getApplicationSandboxConfigurationDefault());
         logger.debug("Returning the current sandbox configurations with {} sandboxes configured.", applicationSandboxConfigurations.getSandboxes().size());
 
         return applicationSandboxConfigurations;
