+++ admin-service/src/test/java/org/opentestsystem/rdw/admin/service/impl/DefaultSandboxServiceTest.java
@@ -43,7 +43,7 @@ public class DefaultSandboxServiceTest {
         verify(inMemorySandboxConfigurationRepository).getAll();
 
         assertThat(applicationSandboxConfigurations).isNotNull();
-        assertThat(applicationSandboxConfigurations.getApplicationSandboxConfiguration()).isNotNull();
+        assertThat(applicationSandboxConfigurations.getSandboxConfigurationPackage()).isNotNull();
         assertThat(applicationSandboxConfigurations.getSandboxes()).isNotEmpty();
         assertThat(applicationSandboxConfigurations.getSandboxes().size()).isEqualTo(2);
 
@@ -125,10 +125,11 @@ public class DefaultSandboxServiceTest {
         assertThat(sandboxConfiguration.getSandbox().getDescription()).isEqualTo(expectedSandboxConfiguration.getSandbox().getDescription());
         assertThat(sandboxConfiguration.getSandbox().getName()).isEqualTo(expectedSandboxConfiguration.getSandbox().getName());
 
-        verifyReporting(sandboxConfiguration.getReporting(), expectedSandboxConfiguration.getReporting());
+        verifyReporting(sandboxConfiguration.getApplicationSandboxConfiguration().getReporting(), expectedSandboxConfiguration.getApplicationSandboxConfiguration().getReporting());
         verifyLocalization(sandboxConfiguration.getLocalization(), expectedSandboxConfiguration.getLocalization());
 
-        verifyAggregatorProperties(sandboxConfiguration.getAggregateReportingPropertiesTenant(), expectedSandboxConfiguration.getAggregateReportingPropertiesTenant());
+        verifyAggregatorProperties(sandboxConfiguration.getApplicationSandboxConfiguration().getAggregateReportingPropertiesTenant(),
+                expectedSandboxConfiguration.getApplicationSandboxConfiguration().getAggregateReportingPropertiesTenant());
     }
 
     /**
