+++ admin-service/src/test/java/org/opentestsystem/rdw/admin/web/SandboxControllerTest.java
@@ -46,7 +46,7 @@ public class SandboxControllerTest {
         Set<SandboxConfiguration> sandboxConfigurations = ImmutableSet.of(sandboxConfigurationCa, sandboxConfigurationMi);
         ApplicationSandboxConfigurations applicationSandboxConfigurations = new ApplicationSandboxConfigurations();
         applicationSandboxConfigurations.setSandboxes(sandboxConfigurations);
-        applicationSandboxConfigurations.setApplicationSandboxConfiguration(SandboxConfigurationStubs.getApplicationSandboxConfigurationDefault());
+        applicationSandboxConfigurations.setSandboxConfigurationPackage(SandboxConfigurationStubs.getApplicationSandboxConfigurationDefault());
 
         when(service.getAll()).thenReturn(applicationSandboxConfigurations);
